Add files via upload
It is a zip folder of my project which is a to-do app made with react.It create and categorize tasks related to your website projects. Assign deadlines, priorities, and labels to each task for better organization and clarity.

Prioritize tasks based on urgency, importance, or project dependencies. This feature ensures that you focus on high-priority items and maintain a productive workflow.

Monitor the progress of your tasks in real-time. Visualize your achievements and identify bottlenecks to make necessary adjustments and stay on track towards meeting your website goals.